+++ Kernel/System/Ticket/Event/NotificationEvent/Transport/Email.pm
@@ -615,7 +615,6 @@ sub SecurityOptionsGet {
         # Take just valid keys.
         @SignKeys = grep { $_->{Status} eq 'good' } @SignKeys;
 
-        # get public keys
         @EncryptKeys = $PGPObject->PublicKeySearch(
             Search => $Param{Recipient}->{UserEmail},
         );
@@ -643,12 +642,16 @@ sub SecurityOptionsGet {
             return;
         }
 
-        @SignKeys = $Kernel::OM->Get('Kernel::System::Crypt::SMIME')->PrivateSearch(
+        # Take just valid keys.
+        @SignKeys = $SMIMEObject->PrivateSearch(
             Search => $NotificationSenderEmail,
+            Valid  => 1,
         );
 
-        @EncryptKeys = $Kernel::OM->Get('Kernel::System::Crypt::SMIME')->CertificateSearch(
+        # Take just valid keys.
+        @EncryptKeys = $SMIMEObject->CertificateSearch(
             Search => $Param{Recipient}->{UserEmail},
+            Valid  => 1,
         );
 
         $SecurityOptions{Backend} = 'SMIME';